Updated docs and craked the code for adding seperate tiles even for a bigger tilesets,added sprites replacing logic with sprite tracking(added docs also)

+++ src/features/mapEditor/types/types.ts
@@ -52,7 +52,7 @@ export type Layer = {
    zindex: number
    locked: boolean
    opacity: number
-   data: Uint16Array
+   data: Uint32Array
 }
 export type MapData = {
    width: number
